Move getQueryableFields to utils
